pyproject.toml
[tool.poetry]
name = "shcommerce"
version = "0.1.0"
description = ""
authors = ["windalfin"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.94.1"
uvicorn = "^0.21.1"
pydantic = {extras = ["email", "dotenv"], version = "^1.10.6"}
fastapi-users = "^10.4.1"
fastapi-users-tortoise = "^0.2.0"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.3"}
aerich = "^0.7.1"
saq = {extras = ["hiredis", "web"], version = "^0.9.3"}
hupper = "^1.11"
python-multipart = "^0.0.6"
sentry-sdk = "^1.17.0"
rich = "^13.3.2"
typer = "^0.7.0"
watchgod = "^0.8.2"
Jinja2 = "^3.1.2"
httpx = "^0.23.3"
aioaws = "^0.14"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
factory-boy = "^3.2.1"
pre-commit = "^3.1.1"
mypy = "^1.1.1"
isort = "^5.12.0"
flake8 = "^6.0.0"
ipython = "^8.11.0"
aiosmtpd = "^1.4.4.post2"
honcho = "^1.1.0"

[tool.aerich]
tortoise_orm = "app.db.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
[tool.isort]
profile = "black"

[tool.mypy]
warn_unreachable = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
ignore_missing_imports = true

[tool.ruff]
line-length = 120
select = [
    "E", # https://github.com/charliermarsh/ruff#pycodestylet
    "F", # https://github.com/charliermarsh/ruff#pyflakes
    "UP", # https://github.com/charliermarsh/ruff#pyupgrade
    "N", # https://github.com/charliermarsh/ruff#pep8-naming
    "S", # https://github.com/charliermarsh/ruff#flake8-bandit
    "C", # https://github.com/charliermarsh/ruff#flake8-comprehensions
    "B", # https://github.com/charliermarsh/ruff#flake8-bugbear
    "T" # https://github.com/charliermarsh/ruff#flake8-print
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
